handler functions seperated; build task added
