+++ .gitignore
@@ -1,3 +1,4 @@
+dist
 # Logs
 logs
 *.log
