+++ package.json
@@ -5,6 +5,8 @@
   "main": "index.js",
   "scripts": {
     "start": "npx gulp",
+    "build": "npx gulp build",
+    "prod": "nodemon dist/bin/www.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -25,11 +27,17 @@
     "autoprefixer": "^9.8.4",
     "babel-loader": "^8.1.0",
     "browser-sync": "^2.26.7",
+    "cssnano": "^4.1.10",
+    "del": "^5.1.0",
     "gulp": "^4.0.2",
+    "gulp-babel": "^8.0.0",
     "gulp-imagemin": "^6.2.0",
+    "gulp-minify-html": "^1.0.6",
     "gulp-postcss": "^8.0.0",
+    "gulp-replace": "^1.0.0",
     "gulp-rev": "^9.0.0",
     "gulp-sass": "^4.1.0",
+    "gulp-uglify": "^3.0.2",
     "node-sass": "^4.14.1",
     "nodemon": "^2.0.4",
     "webpack": "^4.43.0",
